+++ crates/mirror-mirror/src/type_info/graph.rs
@@ -1,34 +1,37 @@
 use alloc::borrow::ToOwned;
 use alloc::boxed::Box;
-use alloc::collections::BTreeMap;
 use alloc::string::String;
 use alloc::vec::Vec;
 use core::any::type_name;
 use core::any::TypeId;
-use core::hash::BuildHasher;
+use core::hash::Hash;
+use core::hash::Hasher;
 use core::ops::Deref;
+use kollect::LinearMap;
 
 use super::*;
 use crate::Value;
 use crate::STATIC_RANDOM_STATE;
 
 /// A `TypeGraph`'s node that refers to a specific type via its `TypeId'.
-#[derive(Clone, Copy, Hash, PartialEq, PartialOrd, Ord, Eq, Debug)]
+#[derive(Clone, Copy, PartialEq, PartialOrd, Ord, Eq, Debug)]
 #[cfg_attr(feature = "speedy", derive(speedy::Readable, speedy::Writable))]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct NodeId(u64);
+
+impl Hash for NodeId {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        state.write_u64(self.0);
+    }
+}
 
 impl NodeId {
     fn new<T>() -> Self
     where
         T: 'static,
     {
-        use core::hash::Hash;
-        use core::hash::Hasher;
-
-        let mut hasher = STATIC_RANDOM_STATE.build_hasher();
-        TypeId::of::<T>().hash(&mut hasher);
-        Self(hasher.finish())
+        let hash = STATIC_RANDOM_STATE.hash_one(TypeId::of::<T>());
+        Self(hash)
     }
 }
 
@@ -52,11 +55,17 @@ impl<T> Deref for WithId<T> {
     }
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Eq, Hash)]
+#[derive(Default, Debug, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "speedy", derive(speedy::Readable, speedy::Writable))]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct TypeGraph {
-    pub(super) map: BTreeMap<NodeId, Option<TypeNode>>,
+    pub(super) map: LinearMap<NodeId, Option<TypeNode>>,
+}
+
+impl Hash for TypeGraph {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.map.hash(state);
+    }
 }
 
 impl TypeGraph {
@@ -71,15 +80,23 @@ impl TypeGraph {
         T: DescribeType,
     {
         let id = NodeId::new::<T>();
+        // because we recursively build the graph by passing the graph itself to the build function for this node
+        // before the node is fully built (and therefore before the fully built node is inserted into the map),
+        // if we have any nested types that repeat, we'd usually get an infinite recursion. to stop that, we insert a
+        // marker of `None` to indicate that such a node is already in the process of being built (within the current call stack),
+        // thus we can just return early and not try to build it again.
         match self.map.get(&id) {
-            // the data is already there
+            // node already exists
             Some(Some(_)) => id,
-            // someone else is currently inserting the data
+            // node is in the process of being built in current call stack
             Some(None) => id,
-            // the data isn't there yet
+            // first time this node has been probed, we are responsible for building
             None => {
+                // first insert marker saying we are building this
                 self.map.insert(id, None);
+                // recursively build including any child nodes
                 let info = f(self).into();
+                // insert fully built root node
                 self.map.insert(id, Some(info));
                 id
             }
@@ -127,9 +144,8 @@ impl_from! { Opaque(OpaqueNode) }
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct StructNode {
     pub(super) type_name: String,
-    pub(super) fields: BTreeMap<String, NamedFieldNode>,
-    pub(super) field_names: Box<[String]>,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) fields: LinearMap<String, NamedFieldNode>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
     pub(super) default_value: Option<Value>,
 }
@@ -137,7 +153,7 @@ pub struct StructNode {
 impl StructNode {
     pub fn new<T>(
         fields: &[NamedFieldNode],
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
     ) -> Self
     where
@@ -149,7 +165,6 @@ impl StructNode {
                 .iter()
                 .map(|field| (field.name.clone(), field.clone()))
                 .collect(),
-            field_names: fields.iter().map(|field| field.name.clone()).collect(),
             metadata: map_metadata(metadata),
             docs: map_docs(docs),
             default_value: T::default_value(),
@@ -157,7 +172,7 @@ impl StructNode {
     }
 }
 
-fn map_metadata(metadata: BTreeMap<&'static str, Value>) -> BTreeMap<String, Value> {
+fn map_metadata(metadata: LinearMap<&'static str, Value>) -> LinearMap<String, Value> {
     metadata
         .into_iter()
         .map(|(key, value)| (key.to_owned(), value))
@@ -174,7 +189,7 @@ fn map_docs(docs: &[&'static str]) -> Box<[String]> {
 pub struct TupleStructNode {
     pub(super) type_name: String,
     pub(super) fields: Vec<UnnamedFieldNode>,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
     pub(super) default_value: Option<Value>,
 }
@@ -182,7 +197,7 @@ pub struct TupleStructNode {
 impl TupleStructNode {
     pub fn new<T>(
         fields: &[UnnamedFieldNode],
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
     ) -> Self
     where
@@ -204,7 +219,7 @@ impl TupleStructNode {
 pub struct EnumNode {
     pub(super) type_name: String,
     pub(super) variants: Vec<VariantNode>,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
     pub(super) default_value: Option<Value>,
 }
@@ -212,7 +227,7 @@ pub struct EnumNode {
 impl EnumNode {
     pub fn new<T>(
         variants: &[VariantNode],
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
     ) -> Self
     where
@@ -242,9 +257,8 @@ pub enum VariantNode {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct StructVariantNode {
     pub(super) name: String,
-    pub(super) fields: BTreeMap<String, NamedFieldNode>,
-    pub(super) field_names: Box<[String]>,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) fields: LinearMap<String, NamedFieldNode>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
 }
 
@@ -252,7 +266,7 @@ impl StructVariantNode {
     pub fn new(
         name: &'static str,
         fields: &[NamedFieldNode],
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
     ) -> Self {
         Self {
@@ -261,7 +275,6 @@ impl StructVariantNode {
                 .iter()
                 .map(|field| (field.name.clone(), field.clone()))
                 .collect(),
-            field_names: fields.iter().map(|field| field.name.clone()).collect(),
             metadata: map_metadata(metadata),
             docs: map_docs(docs),
         }
@@ -274,7 +287,7 @@ impl StructVariantNode {
 pub struct TupleVariantNode {
     pub(super) name: String,
     pub(super) fields: Vec<UnnamedFieldNode>,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
 }
 
@@ -282,7 +295,7 @@ impl TupleVariantNode {
     pub fn new(
         name: &'static str,
         fields: &[UnnamedFieldNode],
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
     ) -> Self {
         Self {
@@ -299,14 +312,14 @@ impl TupleVariantNode {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct UnitVariantNode {
     pub(super) name: String,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
 }
 
 impl UnitVariantNode {
     pub fn new(
         name: &'static str,
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
     ) -> Self {
         Self {
@@ -323,14 +336,14 @@ impl UnitVariantNode {
 pub struct TupleNode {
     pub(super) type_name: String,
     pub(super) fields: Vec<UnnamedFieldNode>,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
 }
 
 impl TupleNode {
     pub fn new<T>(
         fields: &[UnnamedFieldNode],
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
     ) -> Self
     where
@@ -351,14 +364,14 @@ impl TupleNode {
 pub struct NamedFieldNode {
     pub(super) name: String,
     pub(super) id: NodeId,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
 }
 
 impl NamedFieldNode {
     pub fn new<T>(
         name: &'static str,
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
         graph: &mut TypeGraph,
     ) -> Self
@@ -379,13 +392,13 @@ impl NamedFieldNode {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct UnnamedFieldNode {
     pub(super) id: NodeId,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) docs: Box<[String]>,
 }
 
 impl UnnamedFieldNode {
     pub fn new<T>(
-        metadata: BTreeMap<&'static str, Value>,
+        metadata: LinearMap<&'static str, Value>,
         docs: &[&'static str],
         graph: &mut TypeGraph,
     ) -> Self
@@ -515,12 +528,12 @@ scalar_typed! {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct OpaqueNode {
     pub(super) type_name: String,
-    pub(super) metadata: BTreeMap<String, Value>,
+    pub(super) metadata: LinearMap<String, Value>,
     pub(super) default_value: Option<Value>,
 }
 
 impl OpaqueNode {
-    pub fn new<T>(metadata: BTreeMap<&'static str, Value>, _graph: &mut TypeGraph) -> Self
+    pub fn new<T>(metadata: LinearMap<&'static str, Value>, _graph: &mut TypeGraph) -> Self
     where
         T: DescribeType,
     {